+++ src/Controller/Admin/AbstractTenantAwareCrudController.php
@@ -18,7 +18,7 @@ abstract class AbstractTenantAwareCrudController extends AbstractCrudController
     }
 
     /** @phpstan-ignore missingType.parameter */
-    private function setTenant($entity): void
+    protected function setTenant($entity): void
     {
         if ($entity instanceof TenantScopedEntityInterface) {
             $user = $this->getUser();