+++ src/Security/TenantScope/Doctrine/TenantFilter.php
@@ -8,6 +8,7 @@ use App\Entity\Interfaces\TenantScopedEntityInterface;
 use App\Security\TenantScope\TenantScopeException;
 use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\Query\Filter\SQLFilter;
+use Symfony\Contracts\Service\Attribute\Required;
 
 /**
  * Class TenantFilter.
@@ -24,6 +25,10 @@ class TenantFilter extends SQLFilter
     {
         // Check if the entity implements the required interfaces
         if (!$targetEntity->getReflectionClass()->implementsInterface(TenantScopedEntityInterface::class)) {
+            return '';
+        }
+
+        if (!$this->hasParameter('tenant_id')) {
             return '';
         }